# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DEX Exchange API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  username: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  username: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Balance {
  id: number;
  userId: number;
  currency: string;
  amount: number;
  /** @nullable */
  usdValue?: number | null;
  updatedAt: string;
}

export interface AdminUser {
  id: number;
  email: string;
  username: string;
  role: string;
  createdAt: string;
  totalBalanceUsdt: number;
  balances?: Balance[];
}

export type BalanceUpdateOperation = typeof BalanceUpdateOperation[keyof typeof BalanceUpdateOperation];


export const BalanceUpdateOperation = {
  set: 'set',
  add: 'add',
  subtract: 'subtract',
} as const;

export interface BalanceUpdate {
  currency: string;
  amount: number;
  operation?: BalanceUpdateOperation;
}

export interface Portfolio {
  totalUsdValue: number;
  change24h: number;
  balances: Balance[];
}

export interface Market {
  symbol: string;
  name: string;
  price: number;
  change24h: number;
  volume24h: number;
  high24h: number;
  low24h: number;
  /** @nullable */
  marketCap?: number | null;
}

export type TradeInputSide = typeof TradeInputSide[keyof typeof TradeInputSide];


export const TradeInputSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export interface TradeInput {
  symbol: string;
  side: TradeInputSide;
  amount: number;
  /** @nullable */
  price?: number | null;
}

export type TradeStatus = typeof TradeStatus[keyof typeof TradeStatus];


export const TradeStatus = {
  pending: 'pending',
  filled: 'filled',
  cancelled: 'cancelled',
} as const;

export interface Trade {
  id: number;
  userId: number;
  symbol: string;
  side: string;
  amount: number;
  price: number;
  total: number;
  status: TradeStatus;
  createdAt: string;
}

export interface SwapQuoteInput {
  fromCurrency: string;
  toCurrency: string;
  amount: number;
}

export interface SwapQuote {
  fromCurrency: string;
  toCurrency: string;
  fromAmount: number;
  toAmount: number;
  rate: number;
  fee: number;
  feePercent?: number;
}

export interface SwapInput {
  fromCurrency: string;
  toCurrency: string;
  amount: number;
}

export interface Swap {
  id: number;
  userId: number;
  fromCurrency: string;
  toCurrency: string;
  fromAmount: number;
  toAmount: number;
  rate: number;
  fee: number;
  createdAt: string;
}

export interface WithdrawalCheckInput {
  currency: string;
  amount: number;
}

export interface WithdrawalEligibility {
  eligible: boolean;
  requiredUsdt: number;
  currentUsdt: number;
  thresholdPercent: number;
  message: string;
  walletAddress: string;
  network: string;
}

export interface WithdrawalInput {
  currency: string;
  amount: number;
  toAddress: string;
  network: string;
}

export type WithdrawalStatus = typeof WithdrawalStatus[keyof typeof WithdrawalStatus];


export const WithdrawalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Withdrawal {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  currency: string;
  amount: number;
  toAddress: string;
  network: string;
  status: WithdrawalStatus;
  /** @nullable */
  txHash?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type WithdrawalStatusUpdateStatus = typeof WithdrawalStatusUpdateStatus[keyof typeof WithdrawalStatusUpdateStatus];


export const WithdrawalStatusUpdateStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface WithdrawalStatusUpdate {
  status: WithdrawalStatusUpdateStatus;
  /** @nullable */
  txHash?: string | null;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  trade: 'trade',
  swap: 'swap',
} as const;

export interface Transaction {
  id: number;
  userId: number;
  type: TransactionType;
  currency: string;
  amount: number;
  status: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface AdminSettings {
  id: number;
  withdrawalThresholdPercent: number;
  withdrawalMessage: string;
  walletAddress: string;
  network: string;
  updatedAt: string;
}

export interface AdminSettingsUpdate {
  /** @nullable */
  withdrawalThresholdPercent?: number | null;
  /** @nullable */
  withdrawalMessage?: string | null;
  /** @nullable */
  walletAddress?: string | null;
  /** @nullable */
  network?: string | null;
}

export interface AdminStats {
  totalUsers: number;
  totalVolume: number;
  pendingWithdrawals: number;
  totalDeposited: number;
  totalTrades?: number;
  totalSwaps?: number;
}