# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DEX Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /swaps:
    post:
      operationId: createSwap
      tags: [swaps]
      summary: Swap one token for another
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwapInput"
      responses:
        "201":
          description: Swap executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Swap"
  /swaps/quote:
    post:
      operationId: getSwapQuote
      tags: [swaps]
      summary: Get a swap quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwapQuoteInput"
      responses:
        "200":
          description: Swap quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapQuote"
  /withdrawals:
    get:
      operationId: getWithdrawals
      tags: [withdrawals]
      summary: Get user withdrawal requests
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Insufficient balance or threshold not met
...
  /admin/users/{userId}/balance:
    patch:
      operationId: adminUpdateBalance
      tags: [admin]
      summary: Add or update crypto balance for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceUpdate"
      responses:
        "200":
          description: Balance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
...
  /admin/withdrawals/{id}:
    patch:
      operationId: adminUpdateWithdrawal
      tags: [admin]
      summary: Approve or reject a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get admin settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    patch:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update admin settings (threshold %, wallet address, network, withdrawal message)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, username, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
...
    Balance:
      type: object
      required: [id, userId, currency, amount, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        currency:
          type: string
        amount:
          type: number
        usdValue:
          type: ["number", "null"]
        updatedAt:
          type: string
    BalanceUpdate:
      type: object
      required: [currency, amount]
      properties:
        currency:
          type: string
        amount:
          type: number
        operation:
          type: string
          enum: [set, add, subtract]
...
    SwapQuoteInput:
      type: object
      required: [fromCurrency, toCurrency, amount]
      properties:
        fromCurrency:
          type: string
        toCurrency:
          type: string
        amount:
          type: number
    SwapQuote:
      type: object
      required: [fromCurrency, toCurrency, fromAmount, toAmount, rate, fee]
      properties:
        fromCurrency:
          type: string
        toCurrency:
          type: string
        fromAmount:
          type: number
        toAmount:
          type: number
        rate:
          type: number
        fee:
          type: number
        feePercent:
          type: number
    SwapInput:
      type: object
      required: [fromCurrency, toCurrency, amount]
      properties:
        fromCurrency:
          type: string
        toCurrency:
          type: string
        amount:
          type: number
    Swap:
      type: object
      required: [id, userId, fromCurrency, toCurrency, fromAmount, toAmount, rate, fee, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        fromCurrency:
          type: string
        toCurrency:
          type: string
        fromAmount:
          type: number
        toAmount:
          type: number
        rate:
          type: number
        fee:
          type: number
        createdAt:
          type: string
...
    WithdrawalInput:
      type: object
      required: [currency, amount, toAddress, network]
      properties:
        currency:
          type: string
        amount:
          type: number
        toAddress:
          type: string
        network:
          type: string
    Withdrawal:
      type: object
      required: [id, userId, currency, amount, toAddress, network, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: ["string", "null"]
        currency:
          type: string
        amount:
          type: number
        toAddress:
          type: string
        network:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        txHash:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    WithdrawalStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        txHash:
          type: ["string", "null"]
...
    AdminSettings:
      type: object
      required: [id, withdrawalThresholdPercent, withdrawalMessage, walletAddress, network, updatedAt]
      properties:
        id:
          type: integer
        withdrawalThresholdPercent:
          type: number
        withdrawalMessage:
          type: string
        walletAddress:
          type: string
        network:
          type: string
        updatedAt:
          type: string
    AdminSettingsUpdate:
      type: object
      properties:
        withdrawalThresholdPercent:
          type: ["number", "null"]
        withdrawalMessage:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
        network:
          type: ["string", "null"]
...