# lib/api-zod/src/generated/types/adminSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DEX Exchange API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSettingsUpdate {
  /** @nullable */
  withdrawalThresholdPercent?: number | null;
  /** @nullable */
  withdrawalMessage?: string | null;
  /** @nullable */
  walletAddress?: string | null;
  /** @nullable */
  network?: string | null;
}