# lib/api-zod/src/generated/types/balance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DEX Exchange API
 * OpenAPI spec version: 0.1.0
 */

export interface Balance {
  id: number;
  userId: number;
  currency: string;
  amount: number;
  /** @nullable */
  usdValue?: number | null;
  updatedAt: string;
}