# lib/api-zod/src/generated/types/withdrawalStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DEX Exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { WithdrawalStatusUpdateStatus } from './withdrawalStatusUpdateStatus';

export interface WithdrawalStatusUpdate {
  status: WithdrawalStatusUpdateStatus;
  /** @nullable */
  txHash?: string | null;
}